% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanExportedMOMOData.R
\name{clean_exported_momo_data}
\alias{clean_exported_momo_data}
\title{CleanExportedMOMOData}
\usage{
clean_exported_momo_data(data, s,
  folder_results = fhi::DashboardFolder("results"))
}
\arguments{
\item{data}{a}

\item{s}{a}

\item{folder_results}{a}
}
\description{
CleanExportedMOMOData
}
