% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_quasipoisson.R
\name{std_save_results_as_excel}
\alias{std_save_results_as_excel}
\title{std_save_results_as_excel}
\usage{

  std_save_results_as_excel(folder_results = fhi::DashboardFolder("results"),
  dateDataMinusOneWeek, data_to_save)
}
\arguments{
\item{folder_results}{a}

\item{dateDataMinusOneWeek}{a}

\item{data_to_save}{a}
}
\description{
std_save_results_as_excel
}
